+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Bug-fixes
 
+- [x] [#536 - Statusbar background color change delayed on application start](https://github.com/scriptPilot/app-framework/issues/536)
 - [x] [#543 - Statusbar shown on initial native app start, but configured as false](https://github.com/scriptPilot/app-framework/issues/543)
 - [x] [#554 - Vue warning due to material icon workaround for Android 4](https://github.com/scriptPilot/app-framework/issues/554)